Some tidy up
