+++ cpp/dolfinx/la/Vector.h
@@ -75,8 +75,14 @@ public:
   {
     const std::int32_t local_size = _bs * _map->size_local();
     xtl::span<const T> x_local(_x.data(), local_size);
-    auto pack = common::Scatterer::pack();
+
+    auto pack = [](const auto& in, const auto& idx, auto& out)
+    {
+      for (std::size_t i = 0; i < idx.size(); ++i)
+        out[i] = in[idx[i]];
+    };
     pack(x_local, _scatterer->local_indices(), _buffer_local);
+
     _scatterer->scatter_fwd_begin(xtl::span<const T>(_buffer_local),
                                   xtl::span<T>(_buffer_remote), _request);
   }
@@ -88,9 +94,14 @@ public:
     const std::int32_t local_size = _bs * _map->size_local();
     const std::int32_t num_ghosts = _bs * _map->num_ghosts();
     xtl::span<T> x_remote(_x.data() + local_size, num_ghosts);
-    auto unpack = common::Scatterer::unpack();
     auto op = [](auto /*a*/, auto b) { return b; };
     _scatterer->scatter_fwd_end(_request);
+
+    auto unpack = [](const auto& in, const auto& idx, auto& out, auto op)
+    {
+      for (std::size_t i = 0; i < idx.size(); ++i)
+        out[idx[i]] = op(out[idx[i]], in[i]);
+    };
     unpack(_buffer_remote, _scatterer->remote_indices(), x_remote, op);
   }
 
@@ -109,8 +120,14 @@ public:
     const std::int32_t local_size = _bs * _map->size_local();
     const std::int32_t num_ghosts = _bs * _map->num_ghosts();
     xtl::span<T> x_remote(_x.data() + local_size, num_ghosts);
-    auto pack = common::Scatterer::pack();
+
+    auto pack = [](const auto& in, const auto& idx, auto& out)
+    {
+      for (std::size_t i = 0; i < idx.size(); ++i)
+        out[i] = in[idx[i]];
+    };
     pack(x_remote, _scatterer->remote_indices(), _buffer_remote);
+
     _scatterer->scatter_rev_begin(xtl::span<const T>(_buffer_remote),
                                   xtl::span<T>(_buffer_local), _request);
   }
@@ -126,8 +143,13 @@ public:
   {
     const std::int32_t local_size = _bs * _map->size_local();
     xtl::span<T> x_local(_x.data(), local_size);
-    auto unpack = common::Scatterer::unpack();
     _scatterer->scatter_rev_end(_request);
+
+    auto unpack = [](const auto& in, const auto& idx, auto& out, auto op)
+    {
+      for (std::size_t i = 0; i < idx.size(); ++i)
+        out[idx[i]] = op(out[idx[i]], in[i]);
+    };
     unpack(_buffer_local, _scatterer->local_indices(), x_local, op);
   }
